+++ lib/palace.dart
@@ -2,7 +2,6 @@ library palace;
 
 /// ? Guards aka middleware s
 export 'package:palace/src/guards/logs_guard.dart';
-export 'package:palace/src/guards/body_parser.dart';
 export 'package:palace/src/guards/cors_guard.dart';
 
 /// ? http