+++ lib/src/guards/body_parser.dart
@@ -1,17 +1,16 @@
-import 'package:http_server/http_server.dart';
+import 'package:palace_body_parser/palace_body_parser.dart' as parser;
 import 'package:palace/palace.dart';
 
 /// will parse the incoming request so if `json` `req.body` type will be `Map` and so on
 class BodyParser {
   void call(Request req, Response res, next) async {
-    /// convert the incoming request body
-    final _reqBody = await HttpBodyHandler.processRequest(req.ioRequest);
-
-    /// it will be Map<String,dynamic>
-    req.body = _reqBody.body;
-
-    /// might be json or form
-    req.bodyType = _reqBody.type;
+    final parsedBody = await parser.parseBodyFromStream(
+      req.request,
+      req.request.headers.contentType != null ? parser.MediaType.parse(req.request.headers.contentType.toString()) : null,
+      req.request.uri,
+      storeOriginalBuffer: false,
+    );
+    req.body = parsedBody.body;
 
     /// call the next guard or handler
     await next();