+++ lib/src/guards/logs_guard.dart
@@ -3,7 +3,7 @@ import 'package:palace/palace.dart';
 class LogsGuard {
   void call(Request req, Response res, Function next) async {
     COLORS.c('''
-    req to => ${req.ioRequest.uri.path},
+    req to => ${req.request.uri.path},
     method => ${req.method}
     ''');
     return await next();