+++ test/guards/body_parser_test.dart
@@ -11,22 +11,21 @@ void main() {
 
   setUp(() async {
     router = Palace();
-    router.use(BodyParser());
     await router.openGates();
   });
   tearDown(() async {
     await router.closeGates();
   });
-  test('if no body then will fail abd respond with bad request', () async {
-    router.all('/home', (req, res) => res.write(req.bodyType!));
-    final res = await _dio.post('/home');
-    print(res.data);
-    expect(res.statusCode, equals(HttpStatus.badRequest));
+  test('if no body then will body will be empty', () async {
+    router.get('/home', (req, res) => res.json(req.body));
+    final res = await _dio.get('/home');
+    expect(res.statusCode, equals(HttpStatus.ok));
+    expect(res.data, equals({}));
   });
   test('every thing is fine with application/json', () async {
-    router.post('/home', (req, res) => res.write(req.bodyType!));
+    router.post('/home', (req, res) => res.write(req.body['name']));
     final res = await _dio.post('/home', options: Options(contentType: 'application/json'), data: {'name': 'queen'});
-    expect(res.data, equals('json'));
+    expect(res.data, equals('queen'));
   });
   test('every thing is fine with application/x-www-form-urlencoded', () async {
     router.post('/home', (req, res) => res.write(req.body['name']));